ytdownload/requirements.txt
Flask==2.3.3
yt-dlp==2023.11.16
python-dotenv==1.0.0
requests==2.31.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
google-auth==2.26.1
Flask-Session==0.5.0
